Keep Supabase emulator integration real
